+++ wombat/src/main/include/PID.h
@@ -93,7 +93,7 @@ namespace wom {
       _integralSum = sum_t{0};
     }
 
-    out_t Calculate(in_t pv, units::second_t dt, out_t feedforward = out_t{0}) {
+    out_t   Calculate(in_t pv, units::second_t dt, out_t feedforward = out_t{0}) {
       auto error = do_wrap(_setpoint - pv);
       _integralSum += error * dt;
       if (config.izone.value() > 0 && (error > config.izone || error < -config.izone))
